+++ pinacles/Containers.py
@@ -235,7 +235,7 @@ class ModelState:
         n = self._Grid.n
 
         local_data = self.get_field(name)[nh[0] : -nh[0], nh[1] : -nh[1], indx]
-        local_copy_of_global = np.zeros((n[0], n[1]), dtype=np.double)
+        local_copy_of_global = np.zeros((n[0], n[1]), dtype=np.single)
 
         local_copy_of_global[ls[0] : ls[0] + nl[0], ls[1] : ls[1] + nl[1]] = local_data
 
@@ -294,28 +294,28 @@ class ModelState:
             if not "ff" in var:  # Avoid SBM Bins
                 if self._loc[var] != "z":
                     v = profiles_grp.createVariable(
-                        var, np.double, dimensions=("time", "z",)
+                        var, np.single, dimensions=("time", "z",)
                     )
                     v.units = self._units[var]
                     v.long_name = self._long_names[var]
                     v.standard_name = "\bar{" + self._latex_names[var] + "}"
 
                     v = profiles_grp.createVariable(
-                        var + "_squared", np.double, dimensions=("time", "z",)
+                        var + "_squared", np.single, dimensions=("time", "z",)
                     )
                     v.units = "{" + self._units[var] + "}^2"
                     v.long_name = self._long_names[var] + " mean of squared"
                     v.standard_name = self._latex_names[var]
 
                     v = profiles_grp.createVariable(
-                        var + "_min", np.double, dimensions=("time", "z",)
+                        var + "_min", np.single, dimensions=("time", "z",)
                     )
                     v.units = self._units[var]
                     v.long_name = "minimum " + self._long_names[var]
                     v.standard_name = "min{" + self._latex_names[var] + "}"
 
                     v = profiles_grp.createVariable(
-                        var + "_max", np.double, dimensions=("time", "z",)
+                        var + "_max", np.single, dimensions=("time", "z",)
                     )
                     v.units = self._units[var]
                     v.long_name = "maximum " + self._long_names[var]
@@ -323,28 +323,28 @@ class ModelState:
 
                 else:
                     v = profiles_grp.createVariable(
-                        var, np.double, dimensions=("time", "z_edge",)
+                        var, np.single, dimensions=("time", "z_edge",)
                     )
                     v.units = self._units[var]
                     v.long_name = self._long_names[var]
                     v.standard_name = "\bar{" + self._latex_names[var] + "}"
 
                     v = profiles_grp.createVariable(
-                        var + "_squared", np.double, dimensions=("time", "z_edge",)
+                        var + "_squared", np.single, dimensions=("time", "z_edge",)
                     )
                     v.units = "{" + self._units[var] + "}^2"
                     v.long_name = self._long_names[var] + " mean of squared"
                     v.standard_name = "\bar{" + self._latex_names[var] + "^2}"
 
                     v = profiles_grp.createVariable(
-                        var + "_min", np.double, dimensions=("time", "z_edge")
+                        var + "_min", np.single, dimensions=("time", "z_edge")
                     )
                     v.units = self._units[var]
                     v.long_name = "minimum " + self._long_names[var]
                     v.standard_name = "min{" + self._latex_names[var] + "}"
 
                     v = profiles_grp.createVariable(
-                        var + "_max", np.double, dimensions=("time", "z_edge")
+                        var + "_max", np.single, dimensions=("time", "z_edge")
                     )
                     v.units = self._units[var]
                     v.long_name = "maximum " + self._long_names[var]
@@ -354,14 +354,14 @@ class ModelState:
         for var in self._dofs:
             if not "ff" in var:  # Avoid SBM Bins
                 v = timeseries_grp.createVariable(
-                    var + "_max", np.double, dimensions=("time",)
+                    var + "_max", np.single, dimensions=("time",)
                 )
                 v.units = self._units[var]
                 v.long_name = "minimum " + self._long_names[var]
                 v.standard_name = "min{" + self._latex_names[var] + "}"
 
                 v = timeseries_grp.createVariable(
-                    var + "_min", np.double, dimensions=("time",)
+                    var + "_min", np.single, dimensions=("time",)
                 )
                 v.units = self._units[var]
                 v.long_name = "maximum " + self._long_names[var]
