+++ pinacles/DumpFields.py
@@ -77,7 +77,7 @@ class DumpFields:
                     "ff" not in v
                 ):  # Exclude the bins from the 3D fields for the same of storage
                     v_nc = rt_grp.createVariable(
-                        v, np.double, dimensions=("time", "X", "Y", "Z")
+                        v, np.single, dimensions=("time", "X", "Y", "Z")
                     )
                     v_nc.units = ac.get_units(v)
                     v_nc.long_names = ac.get_long_name(v)
@@ -101,25 +101,25 @@ class DumpFields:
 
         nhalo = self._Grid.n_halo
 
-        T = rt_grp.createVariable("time", np.double, dimensions=("time"))
+        T = rt_grp.createVariable("time", np.single, dimensions=("time"))
         T[0] = self._TimeSteppingController.time
         T.units = "s"
         T.long_name = "time"
         T.standard_name = "t"
 
-        X = rt_grp.createVariable("X", np.double, dimensions=("X"))
+        X = rt_grp.createVariable("X", np.single, dimensions=("X"))
         X.units = "m"
         X.long_name = "x-coordinate"
         X.standard_name = "x"
         X[:] = self._Grid.x_local[nhalo[0] : -nhalo[0]]
 
-        Y = rt_grp.createVariable("Y", np.double, dimensions=("Y"))
+        Y = rt_grp.createVariable("Y", np.single, dimensions=("Y"))
         Y.units = "m"
         Y.long_name = "y-coordinate"
         Y.standard_name = "y"
         Y[:] = self._Grid.y_local[nhalo[1] : -nhalo[1]]
 
-        Z = rt_grp.createVariable("Z", np.double, dimensions=("Z"))
+        Z = rt_grp.createVariable("Z", np.single, dimensions=("Z"))
         Z.units = "m"
         Z.long_name = "z-coordinate"
         Z.standard_name = "z"
